pyproject.toml
[tool.poetry]
name = "blog-work"
version = "0.1.0"
description = ""
authors = ["Ayumu <hrtaym1114@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
swarm = {git = "https://github.com/openai/swarm.git"}
jupyterlab = "^4.2.5"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pandas = "^2.2.3"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
